+++ network/domain/src/test/kotlin/me/proton/core/network/domain/HumanVerificationHandlerTest.kt
@@ -52,7 +52,7 @@ class HumanVerificationHandlerTest {
 
     @BeforeTest
     fun beforeTest() {
-        every { clientIdProvider.getClientId(any()) } returns clientId
+        coEvery { clientIdProvider.getClientId(any()) } returns clientId
         coEvery { humanVerificationListener.notifyHumanVerificationProcessFinished(any()) } answers {}
         // Assume no token has been refreshed between each tests.
         runBlocking { HumanVerificationNeededHandler.reset(clientId) }